Move AddFutureVisitAsync to DbSeeder
